Hold a compiled program's values in long long, not the host's long
cc1 used the host's 'long' as the type every value belonging to the program
being compiled travels in - a literal, a case label, an array length, a
folded constant, a byte pattern laid down as data. That is correct while
the host has a 64-bit long, which every machine this has ever been built on
does. It is wrong on Windows, where long is four bytes, and the effect is
not a diagnostic: a value simply loses its top half somewhere between the
lexer and the assembler.

143 declarations across the AST, the lexer, the parser, the preprocessor
and all three backends now say long long. Driver.cpp is deliberately
untouched - every long in it is a pid, a vector index or a core id read out
of /proc, none of them a value belonging to the compiled program, and
widening those would say something untrue about what they are.

Three conversions mattered as much as the declarations, and were the reason
the first attempt still failed: strtoul returns unsigned long and strtol
returns long, so the literal was already truncated before anything could
store it. strtoull and strtoll now.

**On Linux and macOS this changes nothing, and that is the evidence.** Both
have a 64-bit long, so every one of these declarations named the same type
before and after: 421 Linux, 8 cross-ABI, 18 Windows-GNU, 16 arm64, and all
412 cases still agree with clang exactly on arm64.

What it changes is a cc1 hosted on Windows, which now compiles the corpus
the way the one built anywhere else does: 405 of 412 agreeing with cl,
against 398 before. The remaining disagreement is ce_unsigned_long_div,
which cl fails too - its arithmetic assumes an LP64 long, and neither
compiler is wrong about it.

Two of the three self-hosted-only failures were literals over 2^32 in
gl_long and ty_mixed_params, and the third was literals_escapes; all three
were the same truncation seen from different sides.

+++ src/Ast.h
@@ -91,16 +91,16 @@ enum class BinOp { Add, Sub, Mul, Div, Mod, Shl, Shr,
 
 class Num final : public Expr {
 public:
-    explicit Num(long v) : value_(v) {}
+    explicit Num(long long v) : value_(v) {}
     // 'long double' rather than double, so that a constant folded at compile
     // time keeps the width the target will give it. Narrowing happens once, in
     // the backend, against the type the expression actually has.
     explicit Num(long double d) : dvalue_(d) {}
-    long value() const { return value_; }
+    long long value() const { return value_; }
     long double dvalue() const { return dvalue_; }
     void accept(Visitor &v) const override { v.visit(*this); }
 private:
-    long value_ = 0;
+    long long value_ = 0;
     long double dvalue_ = 0;
 };
 
@@ -240,16 +240,16 @@ private:
 
 class Postfix final : public Expr {
 public:
-    Postfix(ExprPtr target, bool increment, long step)
+    Postfix(ExprPtr target, bool increment, long long step)
         : target_(std::move(target)), increment_(increment), step_(step) {}
     const Expr &target() const { return *target_; }
     bool increment() const { return increment_; }
-    long step() const { return step_; }
+    long long step() const { return step_; }
     void accept(Visitor &v) const override { v.visit(*this); }
 private:
     ExprPtr target_;
     bool increment_;
-    long step_;
+    long long step_;
 };
 
 class Cast final : public Expr {
@@ -389,15 +389,15 @@ private:
 
 class Case final : public Stmt {
 public:
-    Case(long value, bool isDefault, int id, StmtPtr body)
+    Case(long long value, bool isDefault, int id, StmtPtr body)
         : value_(value), isDefault_(isDefault), id_(id), body_(std::move(body)) {}
-    long value() const { return value_; }
+    long long value() const { return value_; }
     bool isDefault() const { return isDefault_; }
     int id() const { return id_; }
     const Stmt &body() const { return *body_; }
     void accept(Visitor &v) const override { v.visit(*this); }
 private:
-    long value_;
+    long long value_;
     bool isDefault_;
     int id_;
     StmtPtr body_;
@@ -491,7 +491,7 @@ private:
 struct GlobalPiece {
     int offset;
     int size;
-    long value;
+    long long value;
     // When this is not empty the piece is an address constant: the address of
     // 'symbol', plus 'value' bytes. The linker resolves it - the program never
     // computes it, which is what makes it a constant at all. C90 6.5.7 allows
